atmega328p wdt: convert to avr-hal-generic macro
